Put the test-cases in a separate module.
