+++ astar/AStar.ts
@@ -142,100 +142,6 @@ module AStar {
         return path;
     }
 
-    /////////////////////////////////////////////////////////////////////////////////////////////////////
-    // Test cases
-
-    // Creates an example graph and runs AStar on it
-    export function testCase1() {
-
-        // Define graph, with perfect heuristics
-        // Right side (should be visited)
-        var a = new Node("a");
-        var b = new Node("b");
-        var c = new Node("c");
-        var d = new Node("d");
-        var e = new Node("e");
-        // Left side (should not be visited, due to heuristics)
-        var f = new Node("f");
-        var g = new Node("g");
-        var h = new Node("h");
-        var nodes = [a,b,c,d,e,f,g,h];
-        var edges = Edge.createEdges([[a,b,1], [b,c,1], [c,d,1], [a,e,1], [e,d,4], // Right side
-                                            [a,f,0.5], [f,g,1], [g,h,1], [h,f,1]]); // Left side
-
-        initGraph(nodes, edges); // Updates node objects to be a proper graph
-
-        //A simple heuristic function that simply returns the exakt cost of the shortest path of the node
-        var heuristic = function(node) {
-            if(node.label === "a") return 3
-            if(node.label === "b") return 2
-            if(node.label === "c") return 1
-            if(node.label === "d") return 0
-            if(node.label === "e") return 4
-            if(node.label === "f") return 3.5
-            if(node.label === "g") return 4.5
-            if(node.label === "h") return 4.5
-        }
-
-        console.log("Running astar correctness test ... ");
-        var path = astar(a, d, nodes, heuristic);
-        var correctPath = [a,b,c,d];
-        if (!test(arrayEquals(path, correctPath)))
-            console.log("nodes: ",nodes);
-
-        console.log("Running astar heuristics test ... ");
-        if (!test(g.previous===null && h.previous===null)) {
-            console.log("nodes: ",nodes);
-            console.log("g.previous:", g.previous);
-            console.log("h.previous:", h.previous);
-        }
-
-    }
-
-    function test(test: boolean) : boolean {
-        console.log(test? "... passed!" : "... FAILED!");
-        return test;
-    }
-
-    /////////////////////////////////////////////////////////////////////////////////////////////////////
-    // Utility functions
-
-    // Creates a graph from a list of blank nodes and edges.
-    // More specifically, updates all nodes by adding the neighbour references/costs specified in edges
-    function initGraph(nodes : Node[], edges : Edge[]) {
-        for (var eKey in edges) {
-            var e1 = edges[eKey];
-            var e2 = e1.complement(); // create opposite edge
-            var v1 = e1.start, v2 = e1.end, c = e1.cost;
-            v1.neighbours.push(e1);
-            v2.neighbours.push(e2);
-        }
-    }
-
-    // Can't extend prototype in typescript? :(
-    //Array.prototype.shallowEquals = ...
-
-    // Compares shallowly if two arrays are equal
-    function arrayEquals<T>(first : Array<T>, second : Array<T>) : boolean {
-        if (!first || !second)              return false;
-        if (first.length !== second.length) return false;
-
-        // Compare all refs in array
-        for (var i=0;i<first.length;i++) {
-            if (first[i] !== second[i]) return false;
-        }
-        return true;
-    }
-
-    // (Not used)
-    function listMinus(a : Object[], b : Object[]) : Object[] {
-        var newA = a.slice(0);
-        return newA.filter((o) => {
-            return b.indexOf(o) !== -1;
-        });
-    }
-
-
     /**
      * This class is copied from collections. We only did this to be able to provide the
      * update function!
