+++ astar/index.html
@@ -3,11 +3,12 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/lodash.js/3.5.0/lodash.min.js"></script>
     <script src="collections.js"></script>
     <script src="AStar.js"></script>
+    <script src="TestCases.js"></script>
 </head>
 <body>
 <script type="text/javascript" language="javascript">
 	console.log("testCase1: ");
-	AStar.testCase1();
+	AStarTestCases.testCase1();
 </script>
 
 </body>
