+++ lib/collections.js
@@ -2312,4 +2312,3 @@ var collections;
     })();
     collections.BSTree = BSTree; // end of BSTree
 })(collections || (collections = {})); // End of module 
-//# sourceMappingURL=collections.js.map
